Move todo contract into control plane todos
