+++ loopx/agent_registry.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from pathlib import Path
 from typing import Any
 
-from .todo_contract import normalize_todo_claimed_by
+from .control_plane.todos.contract import normalize_todo_claimed_by
 
 
 def normalize_registered_agents(values: Any) -> list[str]:
