+++ loopx/control_plane/agents/agent_scope.py
@@ -6,7 +6,7 @@ from typing import Any
 
 from ...decision_scope import todo_gate_relation, todo_gate_relation_blocks_agent
 from ...policies.work_lane import work_lane_contract_requires_current_agent_attempt
-from ...todo_contract import (
+from ..todos.contract import (
     TODO_STATUS_OPEN,
     TODO_TASK_CLASS_ADVANCEMENT,
     TODO_TASK_CLASS_MONITOR,
