+++ loopx/decision_scope.py
@@ -9,7 +9,7 @@ from __future__ import annotations
 
 from typing import Any
 
-from .todo_contract import (
+from .control_plane.todos.contract import (
     normalize_todo_decision_scope,
     normalize_todo_id,
     normalize_todo_required_decision_scopes,
