+++ loopx/status.py
@@ -331,7 +331,7 @@ from .state_projection import (
     next_action_projection_warning,
     state_projection_gap_warning,
 )
-from .todo_contract import (
+from .control_plane.todos.contract import (
     TODO_STATUS_OPEN,
     TODO_TASK_CLASS_ADVANCEMENT,
     TODO_TASK_CLASS_MONITOR,
