+++ loopx/task_lease.py
@@ -9,7 +9,7 @@ from typing import Any
 from .file_lock import exclusive_file_lock
 from .history import load_registry
 from .paths import resolve_runtime_root
-from .todo_contract import (
+from .control_plane.todos.contract import (
     normalize_required_write_scopes,
     normalize_todo_claimed_by,
     normalize_todo_id,
